Added user.js + cleaned up users.js
Added a new file `user.js`. This file is used for functions related to a single user. These include esi `/ui/` functions as well as saving a new refresh token and `/location/`.

The general idea is that only functions that affect multiple users will be in `users.js`.
